6:29 p.m.  I have to put it down and walk away, so I can plan the next attack on the <type_mark> class and finish out the <procedure_header> .
I have to make sure I know what is required and what is not required, so when something is required by the grammar, I can use the token-counter option to make sure the next token read is what is expected.  If the token is optional, I need to use the isValid() method that I have created in each of the classes to delete the newly created object from the parse tree if it is not needed.  For example, if there are no declarations in the program, there is no need to have procedure declarations in the tree and there is no need to have a <declaration> type either.  However, if there are no declarations and the user adds the reserved word "procedure", we have a problem.